Added full flow PS1 GSB CSO test
