+++ frontend/e2e-tests/test-data/specific-data-entries/AB2023_Limburg.ts
@@ -1,0 +1,567 @@
+import type {
+  CommonPollingStationResults,
+  CSOFirstSessionResults,
+  DSOFirstSessionResults,
+} from "@/types/generated/openapi";
+
+export const commonDataEntryAB2023_Limburg: CommonPollingStationResults = {
+  voters_counts: {
+    poll_card_count: 6960,
+    proxy_certificate_count: 803,
+    voter_card_count: 3,
+    total_admitted_voters_count: 7766,
+  },
+  votes_counts: {
+    political_group_total_votes: [
+      { number: 1, total: 56 },
+      { number: 2, total: 776 },
+      { number: 3, total: 12 },
+      { number: 4, total: 143 },
+      { number: 5, total: 510 },
+      { number: 6, total: 14 },
+      { number: 7, total: 578 },
+      { number: 8, total: 53 },
+      { number: 9, total: 6 },
+      { number: 10, total: 9 },
+      { number: 11, total: 3092 },
+      { number: 12, total: 7 },
+      { number: 13, total: 96 },
+      { number: 14, total: 174 },
+      { number: 15, total: 565 },
+      { number: 16, total: 43 },
+      { number: 17, total: 1275 },
+      { number: 18, total: 196 },
+    ],
+    total_votes_candidates_count: 7605,
+    blank_votes_count: 129,
+    invalid_votes_count: 33,
+    total_votes_cast_count: 7767,
+  },
+  differences_counts: {
+    more_ballots_count: 1,
+    fewer_ballots_count: 0,
+    compare_votes_cast_admitted_voters: {
+      admitted_voters_equal_votes_cast: false,
+      votes_cast_greater_than_admitted_voters: true,
+      votes_cast_smaller_than_admitted_voters: false,
+    },
+    difference_completely_accounted_for: { yes: false, no: true },
+  },
+  political_group_votes: [
+    {
+      number: 1,
+      total: 56,
+      candidate_votes: [
+        { number: 1, votes: 42 },
+        { number: 2, votes: 0 },
+        { number: 3, votes: 1 },
+        { number: 4, votes: 3 },
+        { number: 5, votes: 3 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 2 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 2 },
+      ],
+    },
+    {
+      number: 2,
+      total: 776,
+      candidate_votes: [
+        { number: 1, votes: 304 },
+        { number: 2, votes: 195 },
+        { number: 3, votes: 50 },
+        { number: 4, votes: 20 },
+        { number: 5, votes: 28 },
+        { number: 6, votes: 4 },
+        { number: 7, votes: 28 },
+        { number: 8, votes: 3 },
+        { number: 9, votes: 5 },
+        { number: 10, votes: 9 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 2 },
+        { number: 13, votes: 1 },
+        { number: 14, votes: 0 },
+        { number: 15, votes: 3 },
+        { number: 16, votes: 2 },
+        { number: 17, votes: 1 },
+        { number: 18, votes: 4 },
+        { number: 19, votes: 1 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 16 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 2 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 2 },
+        { number: 26, votes: 1 },
+        { number: 27, votes: 73 },
+        { number: 28, votes: 1 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 0 },
+        { number: 31, votes: 4 },
+        { number: 32, votes: 3 },
+        { number: 33, votes: 1 },
+        { number: 34, votes: 0 },
+        { number: 35, votes: 2 },
+        { number: 36, votes: 1 },
+        { number: 37, votes: 0 },
+        { number: 38, votes: 2 },
+        { number: 39, votes: 0 },
+        { number: 40, votes: 2 },
+        { number: 41, votes: 0 },
+        { number: 42, votes: 0 },
+        { number: 43, votes: 0 },
+        { number: 44, votes: 0 },
+        { number: 45, votes: 2 },
+        { number: 46, votes: 0 },
+        { number: 47, votes: 0 },
+        { number: 48, votes: 0 },
+        { number: 49, votes: 1 },
+        { number: 50, votes: 0 },
+      ],
+    },
+    {
+      number: 3,
+      total: 12,
+      candidate_votes: [
+        { number: 1, votes: 7 },
+        { number: 2, votes: 2 },
+        { number: 3, votes: 2 },
+        { number: 4, votes: 1 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 0 },
+        { number: 15, votes: 0 },
+      ],
+    },
+    {
+      number: 4,
+      total: 143,
+      candidate_votes: [
+        { number: 1, votes: 107 },
+        { number: 2, votes: 3 },
+        { number: 3, votes: 6 },
+        { number: 4, votes: 6 },
+        { number: 5, votes: 6 },
+        { number: 6, votes: 5 },
+        { number: 7, votes: 1 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 2 },
+        { number: 15, votes: 2 },
+      ],
+    },
+    {
+      number: 5,
+      total: 510,
+      candidate_votes: [
+        { number: 1, votes: 244 },
+        { number: 2, votes: 5 },
+        { number: 3, votes: 2 },
+        { number: 4, votes: 11 },
+        { number: 5, votes: 21 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 1 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 2 },
+        { number: 13, votes: 4 },
+        { number: 14, votes: 0 },
+        { number: 15, votes: 3 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 6 },
+        { number: 19, votes: 0 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 48 },
+        { number: 22, votes: 1 },
+        { number: 23, votes: 0 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 0 },
+        { number: 26, votes: 0 },
+        { number: 27, votes: 135 },
+        { number: 28, votes: 1 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 1 },
+        { number: 31, votes: 12 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 2 },
+        { number: 34, votes: 0 },
+        { number: 35, votes: 0 },
+        { number: 36, votes: 0 },
+        { number: 37, votes: 0 },
+        { number: 38, votes: 1 },
+        { number: 39, votes: 0 },
+        { number: 40, votes: 2 },
+        { number: 41, votes: 0 },
+        { number: 42, votes: 4 },
+      ],
+    },
+    {
+      number: 6,
+      total: 14,
+      candidate_votes: [
+        { number: 1, votes: 8 },
+        { number: 2, votes: 0 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 3 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 0 },
+      ],
+    },
+    {
+      number: 7,
+      total: 578,
+      candidate_votes: [
+        { number: 1, votes: 364 },
+        { number: 2, votes: 76 },
+        { number: 3, votes: 20 },
+        { number: 4, votes: 15 },
+        { number: 5, votes: 2 },
+        { number: 6, votes: 9 },
+        { number: 7, votes: 5 },
+        { number: 8, votes: 21 },
+        { number: 9, votes: 4 },
+        { number: 10, votes: 5 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 2 },
+        { number: 13, votes: 3 },
+        { number: 14, votes: 21 },
+        { number: 15, votes: 1 },
+        { number: 16, votes: 3 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 3 },
+        { number: 19, votes: 2 },
+        { number: 20, votes: 4 },
+        { number: 21, votes: 3 },
+        { number: 22, votes: 0 },
+        { number: 23, votes: 3 },
+        { number: 24, votes: 1 },
+        { number: 25, votes: 0 },
+        { number: 26, votes: 1 },
+        { number: 27, votes: 1 },
+        { number: 28, votes: 5 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 0 },
+        { number: 31, votes: 1 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 0 },
+        { number: 34, votes: 2 },
+      ],
+    },
+    {
+      number: 8,
+      total: 53,
+      candidate_votes: [
+        { number: 1, votes: 28 },
+        { number: 2, votes: 2 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 1 },
+        { number: 5, votes: 2 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 2 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 2 },
+        { number: 19, votes: 9 },
+        { number: 20, votes: 1 },
+      ],
+    },
+    {
+      number: 9,
+      total: 6,
+      candidate_votes: [
+        { number: 1, votes: 5 },
+        { number: 2, votes: 0 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 0 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 1 },
+      ],
+    },
+    {
+      number: 10,
+      total: 9,
+      candidate_votes: [
+        { number: 1, votes: 7 },
+        { number: 2, votes: 1 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 0 },
+        { number: 5, votes: 1 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+      ],
+    },
+    {
+      number: 11,
+      total: 3092,
+      candidate_votes: [
+        { number: 1, votes: 355 },
+        { number: 2, votes: 139 },
+        { number: 3, votes: 10 },
+        { number: 4, votes: 1143 },
+        { number: 5, votes: 45 },
+        { number: 6, votes: 95 },
+        { number: 7, votes: 8 },
+        { number: 8, votes: 427 },
+        { number: 9, votes: 3 },
+        { number: 10, votes: 12 },
+        { number: 11, votes: 11 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 20 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 408 },
+        { number: 16, votes: 6 },
+        { number: 17, votes: 386 },
+        { number: 18, votes: 22 },
+      ],
+    },
+    {
+      number: 12,
+      total: 7,
+      candidate_votes: [
+        { number: 1, votes: 4 },
+        { number: 2, votes: 1 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 0 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 1 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 0 },
+      ],
+    },
+    {
+      number: 13,
+      total: 96,
+      candidate_votes: [
+        { number: 1, votes: 76 },
+        { number: 2, votes: 2 },
+        { number: 3, votes: 6 },
+        { number: 4, votes: 1 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 1 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 0 },
+        { number: 19, votes: 0 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 0 },
+        { number: 22, votes: 0 },
+        { number: 23, votes: 1 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 0 },
+        { number: 26, votes: 0 },
+        { number: 27, votes: 1 },
+        { number: 28, votes: 1 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 1 },
+        { number: 31, votes: 0 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 0 },
+        { number: 34, votes: 0 },
+        { number: 35, votes: 0 },
+        { number: 36, votes: 0 },
+        { number: 37, votes: 0 },
+        { number: 38, votes: 0 },
+        { number: 39, votes: 0 },
+        { number: 40, votes: 0 },
+        { number: 41, votes: 0 },
+        { number: 42, votes: 0 },
+        { number: 43, votes: 0 },
+        { number: 44, votes: 0 },
+        { number: 45, votes: 0 },
+        { number: 46, votes: 0 },
+        { number: 47, votes: 0 },
+        { number: 48, votes: 0 },
+        { number: 49, votes: 1 },
+      ],
+    },
+    {
+      number: 14,
+      total: 174,
+      candidate_votes: [
+        { number: 1, votes: 109 },
+        { number: 2, votes: 32 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 12 },
+        { number: 5, votes: 3 },
+        { number: 6, votes: 3 },
+        { number: 7, votes: 13 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 1 },
+      ],
+    },
+    {
+      number: 15,
+      total: 565,
+      candidate_votes: [
+        { number: 1, votes: 385 },
+        { number: 2, votes: 51 },
+        { number: 3, votes: 75 },
+        { number: 4, votes: 2 },
+        { number: 5, votes: 13 },
+        { number: 6, votes: 4 },
+        { number: 7, votes: 6 },
+        { number: 8, votes: 2 },
+        { number: 9, votes: 3 },
+        { number: 10, votes: 7 },
+        { number: 11, votes: 17 },
+      ],
+    },
+    {
+      number: 16,
+      total: 43,
+      candidate_votes: [
+        { number: 1, votes: 27 },
+        { number: 2, votes: 5 },
+        { number: 3, votes: 3 },
+        { number: 4, votes: 2 },
+        { number: 5, votes: 1 },
+        { number: 6, votes: 1 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 2 },
+      ],
+    },
+    {
+      number: 17,
+      total: 1275,
+      candidate_votes: [
+        { number: 1, votes: 517 },
+        { number: 2, votes: 49 },
+        { number: 3, votes: 4 },
+        { number: 4, votes: 13 },
+        { number: 5, votes: 16 },
+        { number: 6, votes: 7 },
+        { number: 7, votes: 24 },
+        { number: 8, votes: 2 },
+        { number: 9, votes: 595 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 4 },
+        { number: 12, votes: 3 },
+        { number: 13, votes: 3 },
+        { number: 14, votes: 3 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 3 },
+        { number: 17, votes: 1 },
+        { number: 18, votes: 3 },
+        { number: 19, votes: 0 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 10 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 8 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 2 },
+        { number: 26, votes: 3 },
+      ],
+    },
+    {
+      number: 18,
+      total: 196,
+      candidate_votes: [
+        { number: 1, votes: 160 },
+        { number: 2, votes: 3 },
+        { number: 3, votes: 4 },
+        { number: 4, votes: 0 },
+        { number: 5, votes: 10 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 1 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 5 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 3 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 0 },
+        { number: 19, votes: 4 },
+      ],
+    },
+  ],
+};
+
+export const dataEntryAB2023_LimburgCSO: CSOFirstSessionResults & { model: "CSOFirstSession" } = {
+  model: "CSOFirstSession",
+  extra_investigation: {
+    extra_investigation_other_reason: { yes: false, no: true },
+    ballots_recounted_extra_investigation: { yes: false, no: true },
+  },
+  counting_differences_polling_station: {
+    difference_ballots_voters_completely_accounted_for: { yes: true, no: false },
+    difference_ballots_per_list: { yes: false, no: true },
+  },
+  ...structuredClone(commonDataEntryAB2023_Limburg),
+};
+
+export const dataEntryAB2023_LimburgDSO: DSOFirstSessionResults & {
+  model: "DSOFirstSession";
+} = {
+  model: "DSOFirstSession",
+  about_report: {
+    corrigendum_present: "TwoDocuments",
+    checks_and_corrections_present: "PagePresent",
+  },
+  checks_and_corrections: {
+    reason_investigation_own_initiative: { unaccounted_difference: true, other_error: true },
+    corrected_results_own_initiative: { yes: true, no: false },
+    corrected_results_csb_request: { yes: false, no: false },
+  },
+  ...structuredClone(commonDataEntryAB2023_Limburg),
+};
