+++ frontend/e2e-tests/test-data/specific-data-entries/PS2023_Drenthe.ts
@@ -1,0 +1,616 @@
+import type {
+  CommonPollingStationResults,
+  CSOFirstSessionResults,
+  DSOFirstSessionResults,
+} from "@/types/generated/openapi";
+
+export const commonDataEntryPS2023_Drenthe: CommonPollingStationResults = {
+  voters_counts: {
+    poll_card_count: 14172,
+    proxy_certificate_count: 1240,
+    voter_card_count: 4,
+    total_admitted_voters_count: 15416,
+  },
+  votes_counts: {
+    political_group_total_votes: [
+      { number: 1, total: 1660 },
+      { number: 2, total: 330 },
+      { number: 3, total: 1553 },
+      { number: 4, total: 646 },
+      { number: 5, total: 1082 },
+      { number: 6, total: 480 },
+      { number: 7, total: 413 },
+      { number: 8, total: 670 },
+      { number: 9, total: 687 },
+      { number: 10, total: 457 },
+      { number: 11, total: 189 },
+      { number: 12, total: 895 },
+      { number: 13, total: 11 },
+      { number: 14, total: 289 },
+      { number: 15, total: 5225 },
+      { number: 16, total: 37 },
+      { number: 17, total: 529 },
+      { number: 18, total: 53 },
+      { number: 19, total: 160 },
+    ],
+    total_votes_candidates_count: 15366,
+    blank_votes_count: 32,
+    invalid_votes_count: 22,
+    total_votes_cast_count: 15420,
+  },
+  differences_counts: {
+    more_ballots_count: 4,
+    fewer_ballots_count: 0,
+    compare_votes_cast_admitted_voters: {
+      admitted_voters_equal_votes_cast: false,
+      votes_cast_greater_than_admitted_voters: true,
+      votes_cast_smaller_than_admitted_voters: false,
+    },
+    difference_completely_accounted_for: { yes: false, no: true },
+  },
+  political_group_votes: [
+    {
+      number: 1,
+      total: 1660,
+      candidate_votes: [
+        { number: 1, votes: 1252 },
+        { number: 2, votes: 60 },
+        { number: 3, votes: 13 },
+        { number: 4, votes: 13 },
+        { number: 5, votes: 6 },
+        { number: 6, votes: 13 },
+        { number: 7, votes: 22 },
+        { number: 8, votes: 11 },
+        { number: 9, votes: 18 },
+        { number: 10, votes: 11 },
+        { number: 11, votes: 8 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 2 },
+        { number: 14, votes: 13 },
+        { number: 15, votes: 3 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 1 },
+        { number: 18, votes: 6 },
+        { number: 19, votes: 7 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 3 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 2 },
+        { number: 24, votes: 29 },
+        { number: 25, votes: 1 },
+        { number: 26, votes: 1 },
+        { number: 27, votes: 3 },
+        { number: 28, votes: 19 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 14 },
+        { number: 31, votes: 3 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 77 },
+        { number: 34, votes: 0 },
+        { number: 35, votes: 2 },
+        { number: 36, votes: 0 },
+        { number: 37, votes: 0 },
+        { number: 38, votes: 2 },
+        { number: 39, votes: 4 },
+        { number: 40, votes: 0 },
+        { number: 41, votes: 0 },
+        { number: 42, votes: 2 },
+        { number: 43, votes: 10 },
+        { number: 44, votes: 2 },
+        { number: 45, votes: 9 },
+        { number: 46, votes: 1 },
+        { number: 47, votes: 1 },
+        { number: 48, votes: 3 },
+        { number: 49, votes: 10 },
+      ],
+    },
+    {
+      number: 2,
+      total: 330,
+      candidate_votes: [
+        { number: 1, votes: 189 },
+        { number: 2, votes: 29 },
+        { number: 3, votes: 11 },
+        { number: 4, votes: 4 },
+        { number: 5, votes: 2 },
+        { number: 6, votes: 3 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 3 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 69 },
+        { number: 15, votes: 1 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 2 },
+        { number: 19, votes: 4 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 0 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 1 },
+        { number: 24, votes: 1 },
+        { number: 25, votes: 4 },
+      ],
+    },
+    {
+      number: 3,
+      total: 1553,
+      candidate_votes: [
+        { number: 1, votes: 971 },
+        { number: 2, votes: 66 },
+        { number: 3, votes: 24 },
+        { number: 4, votes: 38 },
+        { number: 5, votes: 11 },
+        { number: 6, votes: 190 },
+        { number: 7, votes: 8 },
+        { number: 8, votes: 6 },
+        { number: 9, votes: 6 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 4 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 6 },
+        { number: 14, votes: 0 },
+        { number: 15, votes: 3 },
+        { number: 16, votes: 2 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 4 },
+        { number: 19, votes: 8 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 27 },
+        { number: 22, votes: 4 },
+        { number: 23, votes: 0 },
+        { number: 24, votes: 23 },
+        { number: 25, votes: 1 },
+        { number: 26, votes: 69 },
+        { number: 27, votes: 0 },
+        { number: 28, votes: 1 },
+        { number: 29, votes: 2 },
+        { number: 30, votes: 0 },
+        { number: 31, votes: 2 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 23 },
+        { number: 34, votes: 1 },
+        { number: 35, votes: 0 },
+        { number: 36, votes: 1 },
+        { number: 37, votes: 2 },
+        { number: 38, votes: 2 },
+        { number: 39, votes: 1 },
+        { number: 40, votes: 44 },
+      ],
+    },
+    {
+      number: 4,
+      total: 646,
+      candidate_votes: [
+        { number: 1, votes: 400 },
+        { number: 2, votes: 7 },
+        { number: 3, votes: 45 },
+        { number: 4, votes: 126 },
+        { number: 5, votes: 5 },
+        { number: 6, votes: 1 },
+        { number: 7, votes: 10 },
+        { number: 8, votes: 3 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 5 },
+        { number: 11, votes: 5 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 3 },
+        { number: 14, votes: 0 },
+        { number: 15, votes: 0 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 1 },
+        { number: 19, votes: 1 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 22 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 0 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 1 },
+        { number: 26, votes: 0 },
+        { number: 27, votes: 0 },
+        { number: 28, votes: 0 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 3 },
+        { number: 31, votes: 0 },
+        { number: 32, votes: 3 },
+      ],
+    },
+    {
+      number: 5,
+      total: 1082,
+      candidate_votes: [
+        { number: 1, votes: 398 },
+        { number: 2, votes: 382 },
+        { number: 3, votes: 20 },
+        { number: 4, votes: 11 },
+        { number: 5, votes: 22 },
+        { number: 6, votes: 6 },
+        { number: 7, votes: 5 },
+        { number: 8, votes: 30 },
+        { number: 9, votes: 22 },
+        { number: 10, votes: 52 },
+        { number: 11, votes: 52 },
+        { number: 12, votes: 19 },
+        { number: 13, votes: 2 },
+        { number: 14, votes: 4 },
+        { number: 15, votes: 4 },
+        { number: 16, votes: 2 },
+        { number: 17, votes: 1 },
+        { number: 18, votes: 3 },
+        { number: 19, votes: 1 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 0 },
+        { number: 22, votes: 5 },
+        { number: 23, votes: 7 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 3 },
+        { number: 26, votes: 2 },
+        { number: 27, votes: 0 },
+        { number: 28, votes: 3 },
+        { number: 29, votes: 3 },
+        { number: 30, votes: 1 },
+        { number: 31, votes: 3 },
+        { number: 32, votes: 2 },
+        { number: 33, votes: 5 },
+        { number: 34, votes: 1 },
+        { number: 35, votes: 11 },
+      ],
+    },
+    {
+      number: 6,
+      total: 480,
+      candidate_votes: [
+        { number: 1, votes: 261 },
+        { number: 2, votes: 25 },
+        { number: 3, votes: 136 },
+        { number: 4, votes: 27 },
+        { number: 5, votes: 10 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 11 },
+        { number: 8, votes: 8 },
+      ],
+    },
+    {
+      number: 7,
+      total: 413,
+      candidate_votes: [
+        { number: 1, votes: 329 },
+        { number: 2, votes: 5 },
+        { number: 3, votes: 18 },
+        { number: 4, votes: 5 },
+        { number: 5, votes: 7 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 3 },
+        { number: 8, votes: 3 },
+        { number: 9, votes: 10 },
+        { number: 10, votes: 1 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 2 },
+        { number: 13, votes: 1 },
+        { number: 14, votes: 2 },
+        { number: 15, votes: 5 },
+        { number: 16, votes: 3 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 1 },
+        { number: 19, votes: 0 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 7 },
+        { number: 22, votes: 0 },
+        { number: 23, votes: 0 },
+        { number: 24, votes: 4 },
+        { number: 25, votes: 0 },
+        { number: 26, votes: 0 },
+        { number: 27, votes: 4 },
+      ],
+    },
+    {
+      number: 8,
+      total: 670,
+      candidate_votes: [
+        { number: 1, votes: 285 },
+        { number: 2, votes: 238 },
+        { number: 3, votes: 3 },
+        { number: 4, votes: 3 },
+        { number: 5, votes: 4 },
+        { number: 6, votes: 6 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 2 },
+        { number: 9, votes: 16 },
+        { number: 10, votes: 7 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 1 },
+        { number: 14, votes: 38 },
+        { number: 15, votes: 2 },
+        { number: 16, votes: 1 },
+        { number: 17, votes: 0 },
+        { number: 18, votes: 5 },
+        { number: 19, votes: 49 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 6 },
+      ],
+    },
+    {
+      number: 9,
+      total: 687,
+      candidate_votes: [
+        { number: 1, votes: 406 },
+        { number: 2, votes: 45 },
+        { number: 3, votes: 61 },
+        { number: 4, votes: 54 },
+        { number: 5, votes: 8 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 35 },
+        { number: 8, votes: 5 },
+        { number: 9, votes: 6 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 1 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 3 },
+        { number: 16, votes: 2 },
+        { number: 17, votes: 1 },
+        { number: 18, votes: 1 },
+        { number: 19, votes: 3 },
+        { number: 20, votes: 1 },
+        { number: 21, votes: 0 },
+        { number: 22, votes: 30 },
+        { number: 23, votes: 4 },
+        { number: 24, votes: 0 },
+        { number: 25, votes: 1 },
+        { number: 26, votes: 1 },
+        { number: 27, votes: 1 },
+        { number: 28, votes: 0 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 1 },
+        { number: 31, votes: 1 },
+        { number: 32, votes: 0 },
+        { number: 33, votes: 0 },
+        { number: 34, votes: 2 },
+        { number: 35, votes: 1 },
+        { number: 36, votes: 0 },
+        { number: 37, votes: 0 },
+        { number: 38, votes: 0 },
+        { number: 39, votes: 1 },
+        { number: 40, votes: 0 },
+        { number: 41, votes: 0 },
+        { number: 42, votes: 4 },
+        { number: 43, votes: 3 },
+        { number: 44, votes: 0 },
+        { number: 45, votes: 0 },
+        { number: 46, votes: 2 },
+      ],
+    },
+    {
+      number: 10,
+      total: 457,
+      candidate_votes: [
+        { number: 1, votes: 206 },
+        { number: 2, votes: 6 },
+        { number: 3, votes: 16 },
+        { number: 4, votes: 3 },
+        { number: 5, votes: 2 },
+        { number: 6, votes: 10 },
+        { number: 7, votes: 124 },
+        { number: 8, votes: 3 },
+        { number: 9, votes: 34 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 5 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 6 },
+        { number: 14, votes: 23 },
+        { number: 15, votes: 1 },
+        { number: 16, votes: 0 },
+        { number: 17, votes: 2 },
+        { number: 18, votes: 0 },
+        { number: 19, votes: 0 },
+        { number: 20, votes: 0 },
+        { number: 21, votes: 2 },
+        { number: 22, votes: 2 },
+        { number: 23, votes: 2 },
+        { number: 24, votes: 1 },
+        { number: 25, votes: 2 },
+        { number: 26, votes: 0 },
+        { number: 27, votes: 0 },
+        { number: 28, votes: 0 },
+        { number: 29, votes: 0 },
+        { number: 30, votes: 2 },
+        { number: 31, votes: 0 },
+        { number: 32, votes: 1 },
+        { number: 33, votes: 0 },
+        { number: 34, votes: 0 },
+        { number: 35, votes: 2 },
+      ],
+    },
+    {
+      number: 11,
+      total: 189,
+      candidate_votes: [
+        { number: 1, votes: 119 },
+        { number: 2, votes: 4 },
+        { number: 3, votes: 39 },
+        { number: 4, votes: 1 },
+        { number: 5, votes: 1 },
+        { number: 6, votes: 1 },
+        { number: 7, votes: 2 },
+        { number: 8, votes: 11 },
+        { number: 9, votes: 3 },
+        { number: 10, votes: 1 },
+        { number: 11, votes: 0 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 1 },
+        { number: 14, votes: 2 },
+        { number: 15, votes: 1 },
+        { number: 16, votes: 3 },
+      ],
+    },
+    {
+      number: 12,
+      total: 895,
+      candidate_votes: [
+        { number: 1, votes: 668 },
+        { number: 2, votes: 44 },
+        { number: 3, votes: 72 },
+        { number: 4, votes: 27 },
+        { number: 5, votes: 23 },
+        { number: 6, votes: 7 },
+        { number: 7, votes: 44 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 3 },
+        { number: 10, votes: 2 },
+        { number: 11, votes: 2 },
+        { number: 12, votes: 3 },
+      ],
+    },
+    {
+      number: 13,
+      total: 11,
+      candidate_votes: [{ number: 1, votes: 11 }],
+    },
+    {
+      number: 14,
+      total: 289,
+      candidate_votes: [
+        { number: 1, votes: 199 },
+        { number: 2, votes: 10 },
+        { number: 3, votes: 22 },
+        { number: 4, votes: 41 },
+        { number: 5, votes: 6 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 4 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 3 },
+        { number: 10, votes: 1 },
+      ],
+    },
+    {
+      number: 15,
+      total: 5225,
+      candidate_votes: [
+        { number: 1, votes: 3356 },
+        { number: 2, votes: 593 },
+        { number: 3, votes: 75 },
+        { number: 4, votes: 95 },
+        { number: 5, votes: 36 },
+        { number: 6, votes: 62 },
+        { number: 7, votes: 53 },
+        { number: 8, votes: 90 },
+        { number: 9, votes: 18 },
+        { number: 10, votes: 156 },
+        { number: 11, votes: 385 },
+        { number: 12, votes: 5 },
+        { number: 13, votes: 15 },
+        { number: 14, votes: 5 },
+        { number: 15, votes: 9 },
+        { number: 16, votes: 52 },
+        { number: 17, votes: 7 },
+        { number: 18, votes: 35 },
+        { number: 19, votes: 27 },
+        { number: 20, votes: 69 },
+        { number: 21, votes: 82 },
+      ],
+    },
+    {
+      number: 16,
+      total: 37,
+      candidate_votes: [
+        { number: 1, votes: 19 },
+        { number: 2, votes: 4 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 10 },
+        { number: 5, votes: 1 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 1 },
+        { number: 12, votes: 0 },
+        { number: 13, votes: 0 },
+        { number: 14, votes: 1 },
+        { number: 15, votes: 1 },
+      ],
+    },
+    {
+      number: 17,
+      total: 529,
+      candidate_votes: [
+        { number: 1, votes: 360 },
+        { number: 2, votes: 11 },
+        { number: 3, votes: 42 },
+        { number: 4, votes: 9 },
+        { number: 5, votes: 11 },
+        { number: 6, votes: 15 },
+        { number: 7, votes: 76 },
+        { number: 8, votes: 5 },
+      ],
+    },
+    {
+      number: 18,
+      total: 53,
+      candidate_votes: [
+        { number: 1, votes: 44 },
+        { number: 2, votes: 6 },
+        { number: 3, votes: 0 },
+        { number: 4, votes: 2 },
+        { number: 5, votes: 0 },
+        { number: 6, votes: 0 },
+        { number: 7, votes: 0 },
+        { number: 8, votes: 1 },
+        { number: 9, votes: 0 },
+        { number: 10, votes: 0 },
+        { number: 11, votes: 0 },
+      ],
+    },
+    {
+      number: 19,
+      total: 160,
+      candidate_votes: [
+        { number: 1, votes: 110 },
+        { number: 2, votes: 10 },
+        { number: 3, votes: 13 },
+        { number: 4, votes: 2 },
+        { number: 5, votes: 11 },
+        { number: 6, votes: 2 },
+        { number: 7, votes: 1 },
+        { number: 8, votes: 0 },
+        { number: 9, votes: 1 },
+        { number: 10, votes: 10 },
+      ],
+    },
+  ],
+};
+
+export const dataEntryPS2023_DrentheCSO: CSOFirstSessionResults & { model: "CSOFirstSession" } = {
+  model: "CSOFirstSession",
+  extra_investigation: {
+    extra_investigation_other_reason: { yes: false, no: true },
+    ballots_recounted_extra_investigation: { yes: false, no: true },
+  },
+  counting_differences_polling_station: {
+    difference_ballots_voters_completely_accounted_for: { yes: true, no: false },
+    difference_ballots_per_list: { yes: false, no: true },
+  },
+  ...structuredClone(commonDataEntryPS2023_Drenthe),
+};
+
+export const dataEntryPS2023_DrentheDSO: DSOFirstSessionResults & {
+  model: "DSOFirstSession";
+} = {
+  model: "DSOFirstSession",
+  about_report: {
+    corrigendum_present: "TwoDocuments",
+    checks_and_corrections_present: "PagePresent",
+  },
+  checks_and_corrections: {
+    reason_investigation_own_initiative: { unaccounted_difference: true, other_error: true },
+    corrected_results_own_initiative: { yes: true, no: false },
+    corrected_results_csb_request: { yes: false, no: false },
+  },
+  ...structuredClone(commonDataEntryPS2023_Drenthe),
+};
