+++ frontend/e2e-tests/tests/full-flow/full-flow-PS1-GSB-CSO.e2e.ts
@@ -42,8 +42,8 @@ import { UserCreateElectionPgObj } from "e2e-tests/page-objects/users/UserCreate
 import { UserCreateRolePgObj } from "e2e-tests/page-objects/users/UserCreateRolePgObj";
 import { UserCreateTypePgObj } from "e2e-tests/page-objects/users/UserCreateTypePgObj";
 import { UserListPgObj } from "e2e-tests/page-objects/users/UserListPgObj";
-import { eml110a_AB2023_Limburg, eml110b_single, eml230b_AB2023_Limburg } from "e2e-tests/test-data/eml-files";
-import { dataEntryAB2023_LimburgCSO } from "e2e-tests/test-data/specific-data-entries/AB2023_Limburg";
+import { eml110a_PS2023_Drenthe, eml110b_single, eml230b_PS2023_Drenthe } from "e2e-tests/test-data/eml-files";
+import { dataEntryPS2023_DrentheCSO } from "e2e-tests/test-data/specific-data-entries/PS2023_Drenthe";
 import type { TestUser } from "e2e-tests/test-data/users";
 import { test } from "../../fixtures";
 
@@ -119,24 +119,24 @@ test.describe("full flow WS GSB CSO", () => {
     const electionsOverviewPage = new ElectionsOverviewPgObj(page);
     await electionsOverviewPage.create.click();
 
-    await uploadElectionAndInputHash(page, eml110a_AB2023_Limburg);
+    await uploadElectionAndInputHash(page, eml110a_PS2023_Drenthe);
 
     const committeeCategoryPage = new CommitteeCategoryPgObj(page);
     await expect(committeeCategoryPage.header).toBeVisible();
     await expect(committeeCategoryPage.gsb).toBeChecked();
     await committeeCategoryPage.next.click();
 
-    await uploadCandidatesAndInputHash(page, eml230b_AB2023_Limburg);
+    await uploadCandidatesAndInputHash(page, eml230b_PS2023_Drenthe);
 
     const selectGSBPage = new SelectGSBPgObj(page);
     await expect(selectGSBPage.header).toBeVisible();
-    await expect(selectGSBPage.regions).toHaveCount(31);
-    await selectGSBPage.clickRegionFromList("0907");
+    await expect(selectGSBPage.regions).toHaveCount(12);
+    await selectGSBPage.clickRegionFromList("1680");
 
     await uploadPollingStations(page, eml110b_single);
 
     const countingMethodPage = new CountingMethodTypePgObj(page);
-    await countingMethodPage.checkHeaderContainsName("Gennep");
+    await countingMethodPage.checkHeaderContainsName("Aa en Hunze");
     await expect(countingMethodPage.cso).not.toBeChecked();
     await expect(countingMethodPage.dso).not.toBeChecked();
     await countingMethodPage.cso.check();
@@ -145,13 +145,13 @@ test.describe("full flow WS GSB CSO", () => {
     const numberOfVotersPage = new NumberOfVotersPgObj(page);
     await expect(numberOfVotersPage.header).toBeVisible();
     await expect(numberOfVotersPage.input).toHaveValue("612694"); // value comes from eml110b
-    await numberOfVotersPage.input.fill("31269");
+    await numberOfVotersPage.input.fill("61269");
     await numberOfVotersPage.next.click();
 
     const checkAndSavePage = new CheckAndSavePgObj(page);
     await expect(checkAndSavePage.header).toBeVisible();
     await expect(checkAndSavePage.committeeCategory).toHaveText("type stembureau: Gemeentelijk stembureau");
-    await expect(checkAndSavePage.numberOfVoters).toHaveText("31.269 kiesgerechtigden");
+    await expect(checkAndSavePage.numberOfVoters).toHaveText("61.269 kiesgerechtigden");
     const election = await checkAndSavePage.saveElection();
 
     electionId = election.id;
@@ -161,7 +161,7 @@ test.describe("full flow WS GSB CSO", () => {
     await electionsOverviewPage.findElectionRowById(electionId).click();
 
     const electionHomePage = new ElectionHome(page);
-    await expect(electionHomePage.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHomePage.header).toHaveText("Provinciale Staten Drenthe 2023");
     const sessionCard = electionHomePage.getCommitteeSessionCard(1);
     await expect(sessionCard).toContainText("Eerste zitting — Klaar voor invoer");
 
@@ -275,15 +275,15 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHome = new ElectionHome(page);
-    await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
     await expect(electionHome.getCommitteeSessionCard(1)).toContainText("Eerste zitting");
     await electionHome.detailsButton.click();
 
     const electionDetails = new ElectionDetailsPgObj(page);
-    await expect(electionDetails.header).toHaveText("Gemeentelijk stembureau Gennep");
+    await expect(electionDetails.header).toHaveText("Gemeentelijk stembureau Aa en Hunze");
     await electionDetails.fillForm("Pannerdam", "18-03-2026", "21:34");
 
-    await expect(electionHome.header).toContainText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toContainText("Provinciale Staten Drenthe 2023");
     await expect(page.getByText("Begon op 18 maart 2026 om 21:34")).toBeVisible();
     await electionHome.startButton.click();
 
@@ -334,7 +334,7 @@ test.describe("full flow WS GSB CSO", () => {
       await expect(dataEntryHomePage.feedback).toContainText(station.name);
       await dataEntryHomePage.start.click();
 
-      await fillDataEntryPagesAndSave(page, dataEntryAB2023_LimburgCSO);
+      await fillDataEntryPagesAndSave(page, dataEntryPS2023_DrentheCSO);
       await expect(dataEntryHomePage.alertDataEntrySaved).toBeVisible();
 
       await logout(page);
@@ -357,7 +357,7 @@ test.describe("full flow WS GSB CSO", () => {
       await expect(dataEntryHomePage.feedback).toContainText(station.name);
       await dataEntryHomePage.start.click();
 
-      await fillDataEntryPagesAndSave(page, dataEntryAB2023_LimburgCSO);
+      await fillDataEntryPagesAndSave(page, dataEntryPS2023_DrentheCSO);
       await expect(dataEntryHomePage.alertDataEntrySaved).toBeVisible();
 
       await logout(page);
@@ -375,7 +375,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHomePage = new ElectionHome(page);
-    await expect(electionHomePage.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHomePage.header).toHaveText("Provinciale Staten Drenthe 2023");
     await electionHomePage.statusButton.click();
 
     const electionStatusPage = new ElectionStatus(page);
@@ -391,7 +391,7 @@ test.describe("full flow WS GSB CSO", () => {
     const download = await downloadPromise;
 
     expect(download.suggestedFilename()).toMatch(
-      /definitieve-documenten_ab2023_gennep_gemeente_gennep-\d{8}-\d{6}.zip/,
+      /definitieve-documenten_ps2023_aaenhunze_gemeente_aa-en-hunze-\d{8}-\d{6}.zip/,
     );
     expect((await stat(await download.path())).size).toBeGreaterThan(1024);
 
@@ -409,7 +409,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHome = new ElectionHome(page);
-    await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
     await electionHome.newSessionButton.click();
     await electionHome.newSessionModalConfirmButton.click();
     await expect(electionHome.getCommitteeSessionCard(2)).toContainText("Tweede zitting");
@@ -428,7 +428,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHomePage = new ElectionHome(page);
-    await expect(electionHomePage.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHomePage.header).toHaveText("Provinciale Staten Drenthe 2023");
 
     const downloadPromise = page.waitForEvent("download");
     await electionHomePage.downloadNa31_2Inlegvel.click();
@@ -451,7 +451,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHome = new ElectionHome(page);
-    await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
     await electionHome.investigationsOverviewButton.click();
 
     const investigationsOverviewPage = new InvestigationOverviewPgObj(page);
@@ -510,7 +510,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHome = new ElectionHome(page);
-    await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
     await electionHome.startDataEntryButton.click();
 
     const electionStatus = new ElectionStatus(page);
@@ -531,7 +531,7 @@ test.describe("full flow WS GSB CSO", () => {
       await overviewPage.findElectionRowById(electionId!).click();
 
       const electionHome = new ElectionHome(page);
-      await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+      await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
       await electionHome.investigationsOverviewButton.click();
 
       const investigationsOverviewPage = new InvestigationOverviewPgObj(page);
@@ -581,11 +581,11 @@ test.describe("full flow WS GSB CSO", () => {
       const listNames = await progressList.allListNames();
 
       const firstCandidatesPage = new CandidatesListPage(page, 0, listNames[0]!);
-      await firstCandidatesPage.fillCandidate(1, 3);
-      await firstCandidatesPage.fillCandidate(3, 0);
+      await firstCandidatesPage.fillCandidate(1, 13);
+      await firstCandidatesPage.fillCandidate(2, 60);
       await firstCandidatesPage.next.click();
 
-      for (let i = 1; i < dataEntryAB2023_LimburgCSO.political_group_votes.length; i++) {
+      for (let i = 1; i < dataEntryPS2023_DrentheCSO.political_group_votes.length; i++) {
         const candidatesPage = new CandidatesListPage(page, i, listNames[i]!);
         await expect(candidatesPage.fieldset).toBeVisible();
         await page.keyboard.press("Shift+Enter");
@@ -621,15 +621,15 @@ test.describe("full flow WS GSB CSO", () => {
       const votersAndVotesPage = new VotersAndVotesPage(page);
       await expect(votersAndVotesPage.fieldset).toBeVisible();
       await votersAndVotesPage.fillInPageAndClickNext(
-        dataEntryAB2023_LimburgCSO.voters_counts,
-        dataEntryAB2023_LimburgCSO.votes_counts,
+        dataEntryPS2023_DrentheCSO.voters_counts,
+        dataEntryPS2023_DrentheCSO.votes_counts,
       );
 
       const differencesPage = new DifferencesPage(page);
       await expect(differencesPage.fieldset).toBeVisible();
-      await differencesPage.fillInPageAndClickNext(dataEntryAB2023_LimburgCSO.differences_counts);
+      await differencesPage.fillInPageAndClickNext(dataEntryPS2023_DrentheCSO.differences_counts);
 
-      await fillCandidatesListPages(page, dataEntryAB2023_LimburgCSO);
+      await fillCandidatesListPages(page, dataEntryPS2023_DrentheCSO);
 
       const checkAndSavePage = new CheckAndSavePage(page);
       await checkAndSavePage.save.click();
@@ -651,7 +651,7 @@ test.describe("full flow WS GSB CSO", () => {
     await overviewPage.findElectionRowById(electionId!).click();
 
     const electionHome = new ElectionHome(page);
-    await expect(electionHome.header).toHaveText("Waterschap Limburg 2023");
+    await expect(electionHome.header).toHaveText("Provinciale Staten Drenthe 2023");
     await expect(electionHome.header).toBeVisible();
     await electionHome.statusButton.click();
 
@@ -663,7 +663,7 @@ test.describe("full flow WS GSB CSO", () => {
     await finishDataEntryPage.finishDataEntry.click();
 
     const electionDetails = new ElectionDetailsPgObj(page);
-    await expect(electionDetails.header).toHaveText("Gemeentelijk stembureau Gennep");
+    await expect(electionDetails.header).toHaveText("Gemeentelijk stembureau Aa en Hunze");
     await electionDetails.locationInput.fill("Pannerdam");
     await electionDetails.dateInput.fill("18-03-2026");
     await electionDetails.timeInput.fill("21:34");
@@ -675,7 +675,7 @@ test.describe("full flow WS GSB CSO", () => {
     await electionHomePage.downloadSecondSessionZip.click();
 
     const download = await downloadPromise;
-    expect(download.suggestedFilename()).toMatch(/correctie_ab2023_gennep_gemeente_gennep-\d{8}-\d{6}.zip/);
+    expect(download.suggestedFilename()).toMatch(/correctie_ps2023_aaenhunze_gemeente_aa-en-hunze-\d{8}-\d{6}.zip/);
     expect((await stat(await download.path())).size).toBeGreaterThan(1024);
 
     await logout(page);
